# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AIALRA Model Router API
  version: 0.1.0
  description: Private Responses-compatible and native Jobs API for quota-aware model routing
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://router.tailnet.example
    description: Private Tailnet endpoint
security:
  - bearerAuth: []
paths:
...
  /api/v1/jobs:
    post:
      operationId: createJob
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job accepted and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Recent jobs
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
  /api/v1/jobs/{id}:
    get:
      operationId: getJob
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Current job snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/jobs/{id}/events:
    get:
      operationId: getJobEvents
      parameters:
        - $ref: "#/components/parameters/JobId"
        - name: after
          in: query
          schema:
            type: integer
            default: -1
        - name: stream
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Event collection or Server-Sent Events stream
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/JobEvent"
            text/event-stream:
              schema:
                type: string
  /api/v1/jobs/{id}/cancel:
    post:
      operationId: cancelJob
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Cancellation requested or terminal job returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /api/v1/jobs/{id}/approvals:
    post:
      operationId: decideJobApproval
      parameters:
        - $ref: "#/components/parameters/JobId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              additionalProperties: false
              properties:
                decision:
                  type: string
                  enum: [approved, denied]
                reason:
                  type: string
                  maxLength: 2000
      responses:
        "201":
          description: Approval decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobEvent"
  /api/v1/batches:
    post:
      operationId: createBatch
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requests]
              properties:
                requests:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Jobs accepted
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Job"
...
  /api/v1/bootstrap/keys:
    post:
      operationId: bootstrapApiKey
      security: []
      parameters:
        - $ref: "#/components/parameters/BootstrapToken"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: First administrator API key, shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/keys:
    get:
      operationId: listApiKeys
      responses:
        "200":
          description: API key metadata without key digests
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ApiKeyMetadata"
    post:
      operationId: createApiKey
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: API key metadata and one-time plaintext key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedApiKey"
  /api/v1/keys/{id}/revoke:
    post:
      operationId: revokeApiKey
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeApiKeyRequest"
      responses:
        "201":
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                required: [id, revoked, alreadyRevoked]
                properties:
                  id:
                    type: string
                    format: uuid
                  revoked:
                    type: boolean
                  alreadyRevoked:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AIALRA_API_KEY
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BootstrapToken:
      name: X-Bootstrap-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
  responses:
...
    Conflict:
      description: Idempotency key reused with a different request digest
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Authentication failed or bootstrap is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    CreateApiKeyRequest:
      type: object
      required: [name, scopes]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [admin, jobs:read, jobs:write, quota:read, keys:write, approvals:write]
        rateLimitPerMinute:
          type: integer
          minimum: 1
          maximum: 10000
          default: 60
        expiresAt:
          type: [string, "null"]
          format: date-time
          description: Omit for the 30-day default; null requests a non-expiring administrator key
    RevokeApiKeyRequest:
      type: object
      required: [confirmationPrefix]
      additionalProperties: false
      properties:
        confirmationPrefix:
          type: string
          minLength: 1
    ApiKeyMetadata:
      type: object
      required: [id, name, prefix, scopes, rateLimitPerMinute, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        rateLimitPerMinute:
          type: integer
        expiresAt:
          type: [string, "null"]
          format: date-time
        revokedAt:
          type: [string, "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: [string, "null"]
          format: date-time
    CreatedApiKey:
      allOf:
        - $ref: "#/components/schemas/ApiKeyMetadata"
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: Plaintext API key returned only once
            replayed:
              type: boolean
              description: True when the same idempotent request is replayed
...
    PermissionProfile:
      type: object
      additionalProperties: false
      properties:
        filesystem:
          type: string
          enum: [none, read, write]
          default: read
        network:
          type: string
          enum: [none, allowlist]
          default: none
        allowedHosts:
          type: array
          maxItems: 32
          items:
            type: string
        requireApprovalForWrites:
          type: boolean
          default: true
        requireApprovalForExternalActions:
          type: boolean
          default: true
    TaskContract:
      type: object
      required: [objective]
      additionalProperties: false
      properties:
        objective:
          type: string
          minLength: 1
          maxLength: 100000
        taskKind:
          type: string
          enum: [bounded, coding, review, planning, batch, general]
          default: general
        requiredContext:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        expectedOutput:
          type: string
        validation:
          type: object
          properties:
            responseSchema:
              type: object
              additionalProperties: true
            acceptanceTests:
              type: array
              items:
                type: string
        dataClassification:
          type: string
          enum: [public, internal, confidential, restricted]
          default: internal
        permissions:
          $ref: "#/components/schemas/PermissionProfile"
        deadlineMs:
          type: integer
          minimum: 1000
          maximum: 3600000
          default: 120000
        budget:
          type: object
          properties:
            maxCodexCredits:
              type: number
              exclusiveMinimum: 0
            maxOutputTokens:
              type: integer
              default: 8192
            maxAttempts:
              type: integer
              minimum: 1
              maximum: 2
              default: 2
        sessionKey:
          type: string
        model:
          type: string
          enum: [auto, luna, terra, sol]
          default: auto
        effort:
          type: string
          enum: [minimal, low, medium, high, xhigh]
          default: medium
        replayable:
          type: boolean
          default: true
        ambiguity:
          type: integer
          minimum: 0
          maximum: 4
        risk:
          type: integer
          minimum: 0
          maximum: 4
    CreateJobRequest:
      type: object
      required: [task]
      additionalProperties: false
      properties:
        task:
          $ref: "#/components/schemas/TaskContract"
        metadata:
          type: object
          additionalProperties:
            type: string
    Job:
      type: object
      required: [id, status, task, usage, createdAt, updatedAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            [
              accepted,
              queued,
              running,
              validating,
              succeeded,
              needs_review,
              failed,
              cancelled,
              expired,
            ]
        task:
          $ref: "#/components/schemas/TaskContract"
        route:
          oneOf:
            - $ref: "#/components/schemas/RouteDecision"
            - type: "null"
        output: {}
        usage:
          $ref: "#/components/schemas/UsageLedger"
        validation: {}
        errorCode:
          type: [string, "null"]
        errorMessage:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    JobEvent:
      type: object
      required: [id, jobId, sequence, type, data, createdAt]
      properties:
        id:
          type: string
        jobId:
          type: string
          format: uuid
        sequence:
          type: integer
        type:
          type: string
          enum: [status, output.delta, tool, approval, validation, usage, error]
        data:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    RouteDecision:
      type: object
      required: [provider, model, effort, policyVersion, reasonCode, sticky]
      properties:
        provider:
          type: string
          const: codex
        model:
          type: string
        effort:
          type: string
        policyVersion:
          type: string
        reasonCode:
          type: string
        sticky:
          const: true
    UsageLedger:
      type: object
      required:
        [
          inputTokens,
          cachedInputTokens,
          outputTokens,
          codexCredits,
          apiEquivalentUsd,
          quotaUsedPercentBefore,
          quotaUsedPercentAfter,
          quotaWindowDeltaPercent,
          allocatedSubscriptionUsd,
        ]
      properties:
        inputTokens:
          type: integer
          minimum: 0
        cachedInputTokens:
          type: integer
          minimum: 0
        outputTokens:
          type: integer
          minimum: 0
        codexCredits:
          type: [number, "null"]
          description: Codex subscription credits calculated from the official credit rate card
        apiEquivalentUsd:
          type: [number, "null"]
          description: Estimated token-only cost at official API rates; not an actual subscription charge
        quotaUsedPercentBefore:
          type: [number, "null"]
          description: Observed quota-window usage before the task
        quotaUsedPercentAfter:
          type: [number, "null"]
          description: Observed quota-window usage after the task
        quotaWindowDeltaPercent:
          type: [number, "null"]
          description: Observed percentage-point change when both snapshots belong to the same quota window
        allocatedSubscriptionUsd:
          type: [number, "null"]
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}